проверьте пожалуйста

+++ lib/features/announcement_screen/calculate_bloc/calculate_event.dart
@@ -5,7 +5,7 @@ abstract class CalculateEvent {}
 
 class GetCalculateEvent extends CalculateEvent {
   final String text;
-  final int daysCount;
+  final double daysCount;
   final int channelId;
   GetCalculateEvent({required this.text, required this.daysCount,required this.channelId});
 }
